seed is now no longer automatic
