+++ database/index.js
@@ -1,6 +1,4 @@
 const mysql = require('mysql');
-const faker = require('faker');
-const fakerator = require('fakerator');
 
 var connection = mysql.createConnection({
   host: 'localhost',
@@ -96,52 +94,10 @@ const patchReview = function(updatedText, newOverallRating, newFoodRating, newSe
   });
 }
 
-function seedData(){
-  //the moment the db connects, we start adding info into restaurants
-  for (let i = 0; i < 100; i++) {
-    let restaurantName = faker.company.companyName();
-    insertIntoRestaurant(restaurantName, (err, results) => {
-      if (err) {
-        throw new Error(err);
-      } else {
-        console.log(`${restaurantName} added successfully`);
-      }
-    });
-  }
-  for (let i = 0; i < 20; i++) {
-    let username = faker.internet.userName();
-    let city = faker.address.city();
-    let state = faker.address.state();   
-    insertIntoUser(username, city, state, (err, results) => {
-      if (err) {
-        throw new Error(err);
-      } else {
-        console.log(`${username} from ${city}, ${state} added successfully`);
-      }
-    });
-  }
-  for (let i = 0; i < 150; i++) {
-    let restaurant_id = Math.ceil(Math.random() * 100);
-    let user_id = Math.ceil(Math.random() * 20);
-    let review = faker.lorem.lines();
-    let author = faker.internet.userName();
-    let date = faker.date.between('2015-01-01', '2018-10-29'); 
-    let overall_rating = Math.floor(Math.random() * 6);
-    let food_rating = Math.floor(Math.random() * 6);
-    let service_rating = Math.floor(Math.random() * 6);
-    let ambience_rating = Math.floor(Math.random() * 6);
-    let restaurantName = faker.company.companyName();
-    insertIntoReview(restaurant_id, user_id, author, review, date, overall_rating, food_rating, service_rating, ambience_rating, restaurantName, (err, results) => {
-      if (err) {
-        throw new Error(err);
-      } else {
-        console.log('Review added successfully.');
-      }
-    });
-  }
-}
-seedData();
 
+
+module.exports.insertIntoRestaurant = insertIntoRestaurant;
+module.exports.insertIntoUser = insertIntoUser;
 module.exports.getAllReviews = getAllReviews;
 module.exports.insertIntoReview = insertIntoReview;
 module.exports.getIndividualReview = getIndividualReview;
