+++ package.json
@@ -6,7 +6,7 @@
     "test": "jest",
     "build": "webpack --watch",
     "start": "nodemon ./server/index.js",
-    "seed": "node ./database/index.js"
+    "seed": "node ./database/seed.js"
   },
   "keywords": [],
   "author": "",
